+++ python/glide-sync/glide_sync/sync_commands/cluster_commands.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Dict, List, Mapping, Optional, cast
+from typing import Dict, List, Mapping, Optional, Union, cast
 
 from glide_shared.commands.batch import ClusterBatch
 from glide_shared.commands.batch_options import ClusterBatchOptions
@@ -950,6 +950,74 @@ class ClusterCommands(CoreCommands):
         return cast(
             TOK,
             self._execute_command(RequestType.FlushDB, args, route),
+        )
+
+        """
+        Copies the value stored at the `source` to the `destination` key. If `destinationDB`
+        is specified, the value will be copied to the database specified by `destinationDB`,
+        otherwise the current database will be used. When `replace` is True, removes the
+        `destination` key first if it already exists, otherwise performs no action.
+
+        See [valkey.io](https://valkey.io/commands/copy) for more details.
+
+        Args:
+            source (TEncodable): The key to the source value.
+            destination (TEncodable): The key where the value should be copied to.
+            destinationDB (Optional[Union[int, str]]): The alternative logical database index for the destination key.
+            replace (Optional[bool]): If the destination key should be removed before copying the value to it.
+
+        Returns:
+            bool: True if the source was copied. Otherwise, return False.
+
+        Examples:
+            >>> client.set("source", "sheep")
+            >>> client.copy(b"source", b"destination", destinationDB=1)
+                True # Source was copied to DB 1
+            >>> client.select(1)
+            >>> client.get("destination")
+                b"sheep"
+
+        Since: Valkey version 9.0.0.
+        """
+        ...
+
+    def copy(
+        self,
+        source: TEncodable,
+        destination: TEncodable,
+        replace: Optional[bool] = None,
+        destinationDB: Optional[Union[int, str]] = None,
+    ) -> bool:
+        """
+        Copies the value stored at the `source` to the `destination` key.
+
+        This method supports two signatures:
+        1. copy(source, destination, replace=None) - original API
+        2. copy(source, destination, replace=None, destinationDB=None) - new API with DB support
+
+        When `replace` is True, removes the `destination` key first if it already exists,
+        otherwise performs no action.
+
+        See [valkey.io](https://valkey.io/commands/copy) for more details.
+
+        Note:
+            When `destinationDB` is not provided, both `source` and `destination` must map to the same hash slot.
+        """
+        # Handle backward compatibility for positional arguments
+        # If destinationDB is a boolean, it's likely the old API: copy(source, destination, replace)
+        if isinstance(destinationDB, bool):
+            replace = destinationDB
+            destinationDB = None
+
+        # Build command arguments
+        args: List[TEncodable] = [source, destination]
+        if destinationDB is not None:
+            args.extend(["DB", str(destinationDB)])
+        if replace is True:
+            args.append("REPLACE")
+        return cast(
+            bool,
+            self._execute_command(RequestType.Copy, args),
         )
 
     def lolwut(
